+++ src/additional/updateTypes.ts
@@ -1,0 +1,5 @@
+export enum UpdateTypes { 
+    delete = 'delete',
+    modify = 'modify',
+    insert = 'insert'
+}