+++ src/additional/textBuilder.ts
@@ -1,16 +1,19 @@
 import { Range } from "vscode";
 import { ALFileCrawler } from "../al/alFileCrawler";
 import { StringFunctions } from "./stringFunctions";
+import { VarDeclaration } from "../al/varDeclaration";
 
 export module TextBuilder {
     let indent: string = "  ";
     let indentPart: string = "    ";
 
-    export function buildVarDeclaration(range: Range, varName: string, varType: string, local: boolean): string {
+    export function buildVarDeclaration(range: Range, varName: string, varType: string, local: boolean): VarDeclaration {
+        let varDeclaration = new VarDeclaration();
+        let declaration: string = "";
         let indentText: string = "";
-        let varDeclaration: string = "";
         let varStartLineNo = local? ALFileCrawler.findLocalVarSectionStartLineNo() : ALFileCrawler.findGlobalVarSectionStartLineNo();
         let noWhiteSpaces: number;
+        let createVarSection: boolean = false;
         if (varStartLineNo === -1) {
             if (local) {
                 varStartLineNo = ALFileCrawler.findProcedureStartLineNo();
@@ -20,7 +23,7 @@ export module TextBuilder {
                     for(let i = 0; i < noWhiteSpaces; i++) {
                         indentText += " ";
                     }
-                    varDeclaration = indentText + "var" + "\n";
+                    declaration = indentText + "var" + "\n";
                 }
             }
             else {
@@ -30,9 +33,10 @@ export module TextBuilder {
                     for(let i = 0; i < noOfWhiteSpaces; i++) {
                         indentText += " ";
                     }
-                    varDeclaration = indentText + "var" + "\n";
+                    declaration = indentText + "var" + "\n";
                 }
             }
+            createVarSection = true;
         }
         else {
             let varText = ALFileCrawler.getText(varStartLineNo);
@@ -44,11 +48,13 @@ export module TextBuilder {
     
         let varEndLineNo = local? ALFileCrawler.findLocalVarSectionEndLineNo(false, varStartLineNo) : ALFileCrawler.findGlobalVarSectionEndLineNo(varStartLineNo);
         if (varEndLineNo < 0) {
-            return "";
+            return varDeclaration;
         }
     
         // varDeclaration += "        " + varName + ": " + varType + ";";
-        varDeclaration += indentText + indentPart + varName + ": " + varType + ";";
+        declaration += indentText + indentPart + varName + ": " + varType + ";";
+        varDeclaration.declaration = declaration;
+        varDeclaration.createsVarSection = createVarSection;
         return varDeclaration;
     }