+++ src/al/varDeclaration.ts
@@ -1,0 +1,8 @@
+
+export class VarDeclaration {
+    public declaration: string = "";
+    public createsVarSection: boolean = false;
+    constructor() {
+    }
+
+}