+++ CHANGELOG.md
@@ -64,3 +64,7 @@ Bugfixing
 
 - Feature added: variables can also be created as global variable
 - Bug fix: Variable Creation does now also work for AL Files created after opening the project
+
+## [0.2.6]
+
+- Jump to created variable if it was created without object type